requirements.txt
asgiref==3.8.1
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
Django==5.0.4
lxml==5.2.1
pdfminer.six==20231228
pycparser==2.22
python-docx==1.1.2
sqlparse==0.5.0
typing_extensions==4.11.0
